Renamed pam to pambdas
