+++ pambdas/other_stuff.py
