+++ pambdas/series.py
